Form
The form is enclosed by the <form> tag, which defines that the data will be sent to the server. The action="/login" attribute points to the URL where the data will be submitted after clicking the "Log in" button. You can change this URL to suit your application. The method="POST" attribute specifies that the data will be sent using the POST method, which is recommended for login operations to avoid exposing data in the URL.

Inputs:

We use two input fields: one for the username and one for the password. The required attribute ensures that the user must fill in these fields before submitting the form.

Button:

The "Log in" button is defined as <button type="submit">, meaning that clicking it will submit the form.

Server-side Handling:

After the form is submitted, the data will be available on the server (e.g., in PHP, Python, Node.js, etc.) for verification. Appropriate logic should be implemented to check if the provided username and password are correct. In case of a failed login attempt, an appropriate message can be returned, and in case of success, the user can be redirected to the main page or another protected section of the application.